// cvat-ui/ltts_src/reducers/interfaces.ts
export interface Reports {
    Data_list: any;
    initialized: boolean;
    fetching: boolean;
    project: any;
}